+++ packages/react/accordion/src/accordion.tsx
@@ -201,7 +201,7 @@ type AccordionImplContextValue = {
 const [AccordionImplProvider, useAccordionContext] =
   createAccordionContext<AccordionImplContextValue>(ACCORDION_NAME);
 
-type AccordionImplElement = React.ElementRef<typeof Primitive.div>;
+type AccordionImplElement = React.ComponentRef<typeof Primitive.div>;
 type PrimitiveDivProps = React.ComponentPropsWithoutRef<typeof Primitive.div>;
 interface AccordionImplProps extends PrimitiveDivProps {
   /**
@@ -331,7 +331,7 @@ type AccordionItemContextValue = { open?: boolean; disabled?: boolean; triggerId
 const [AccordionItemProvider, useAccordionItemContext] =
   createAccordionContext<AccordionItemContextValue>(ITEM_NAME);
 
-type AccordionItemElement = React.ElementRef<typeof CollapsiblePrimitive.Root>;
+type AccordionItemElement = React.ComponentRef<typeof CollapsiblePrimitive.Root>;
 type CollapsibleProps = React.ComponentPropsWithoutRef<typeof CollapsiblePrimitive.Root>;
 interface AccordionItemProps
   extends Omit<CollapsibleProps, 'open' | 'defaultOpen' | 'onOpenChange'> {
@@ -396,7 +396,7 @@ AccordionItem.displayName = ITEM_NAME;
 
 const HEADER_NAME = 'AccordionHeader';
 
-type AccordionHeaderElement = React.ElementRef<typeof Primitive.h3>;
+type AccordionHeaderElement = React.ComponentRef<typeof Primitive.h3>;
 type PrimitiveHeading3Props = React.ComponentPropsWithoutRef<typeof Primitive.h3>;
 interface AccordionHeaderProps extends PrimitiveHeading3Props {}
 
@@ -429,7 +429,7 @@ AccordionHeader.displayName = HEADER_NAME;
 
 const TRIGGER_NAME = 'AccordionTrigger';
 
-type AccordionTriggerElement = React.ElementRef<typeof CollapsiblePrimitive.Trigger>;
+type AccordionTriggerElement = React.ComponentRef<typeof CollapsiblePrimitive.Trigger>;
 type CollapsibleTriggerProps = React.ComponentPropsWithoutRef<typeof CollapsiblePrimitive.Trigger>;
 interface AccordionTriggerProps extends CollapsibleTriggerProps {}
 
@@ -467,7 +467,7 @@ AccordionTrigger.displayName = TRIGGER_NAME;
 
 const CONTENT_NAME = 'AccordionContent';
 
-type AccordionContentElement = React.ElementRef<typeof CollapsiblePrimitive.Content>;
+type AccordionContentElement = React.ComponentRef<typeof CollapsiblePrimitive.Content>;
 type CollapsibleContentProps = React.ComponentPropsWithoutRef<typeof CollapsiblePrimitive.Content>;
 interface AccordionContentProps extends CollapsibleContentProps {}