+++ packages/react/radio-group/src/radio-group.tsx
@@ -30,14 +30,14 @@ type RadioGroupContextValue = {
   name?: string;
   required: boolean;
   disabled: boolean;
-  value: string;
+  value: string | null;
   onValueChange(value: string): void;
 };
 
 const [RadioGroupProvider, useRadioGroupContext] =
   createRadioGroupContext<RadioGroupContextValue>(RADIO_GROUP_NAME);
 
-type RadioGroupElement = React.ElementRef<typeof Primitive.div>;
+type RadioGroupElement = React.ComponentRef<typeof Primitive.div>;
 type RovingFocusGroupProps = React.ComponentPropsWithoutRef<typeof RovingFocusGroup.Root>;
 type PrimitiveDivProps = React.ComponentPropsWithoutRef<typeof Primitive.div>;
 interface RadioGroupProps extends PrimitiveDivProps {
@@ -48,7 +48,7 @@ interface RadioGroupProps extends PrimitiveDivProps {
   orientation?: RovingFocusGroupProps['orientation'];
   loop?: RovingFocusGroupProps['loop'];
   defaultValue?: string;
-  value?: RadioGroupContextValue['value'];
+  value?: string | null;
   onValueChange?: RadioGroupContextValue['onValueChange'];
 }
 
@@ -71,8 +71,8 @@ const RadioGroup = React.forwardRef<RadioGroupElement, RadioGroupProps>(
     const direction = useDirection(dir);
     const [value, setValue] = useControllableState({
       prop: valueProp,
-      defaultProp: defaultValue ?? '',
-      onChange: onValueChange,
+      defaultProp: defaultValue ?? null,
+      onChange: onValueChange as (value: string | null) => void,
       caller: RADIO_GROUP_NAME,
     });
 
@@ -115,7 +115,7 @@ RadioGroup.displayName = RADIO_GROUP_NAME;
 
 const ITEM_NAME = 'RadioGroupItem';
 
-type RadioGroupItemElement = React.ElementRef<typeof Radio>;
+type RadioGroupItemElement = React.ComponentRef<typeof Radio>;
 type RadioProps = React.ComponentPropsWithoutRef<typeof Radio>;
 interface RadioGroupItemProps extends Omit<RadioProps, 'onCheck' | 'name'> {
   value: string;
@@ -128,7 +128,7 @@ const RadioGroupItem = React.forwardRef<RadioGroupItemElement, RadioGroupItemPro
     const isDisabled = context.disabled || disabled;
     const rovingFocusGroupScope = useRovingFocusGroupScope(__scopeRadioGroup);
     const radioScope = useRadioScope(__scopeRadioGroup);
-    const ref = React.useRef<React.ElementRef<typeof Radio>>(null);
+    const ref = React.useRef<React.ComponentRef<typeof Radio>>(null);
     const composedRefs = useComposedRefs(forwardedRef, ref);
     const checked = context.value === itemProps.value;
     const isArrowKeyPressedRef = React.useRef(false);
@@ -190,7 +190,7 @@ RadioGroupItem.displayName = ITEM_NAME;
 
 const INDICATOR_NAME = 'RadioGroupIndicator';
 
-type RadioGroupIndicatorElement = React.ElementRef<typeof RadioIndicator>;
+type RadioGroupIndicatorElement = React.ComponentRef<typeof RadioIndicator>;
 type RadioIndicatorProps = React.ComponentPropsWithoutRef<typeof RadioIndicator>;
 interface RadioGroupIndicatorProps extends RadioIndicatorProps {}