+++ packages/react/checkbox/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @radix-ui/react-checkbox
 
+## 1.3.2
+
+- Replace deprecated 'ElementRef' with 'ComponentRef' (#3426)
+- Updated dependencies: `@radix-ui/react-primitive@2.1.3`
+
+## 1.3.1
+
+- Fix type error emitted in build artifacts
+- Updated dependencies: `@radix-ui/react-primitive@2.1.2`
+
+## 1.3.0
+
+- Add unstable `Provider`, `Trigger` and `BubbleInput` parts to Checkbox ([#3459](https://github.com/radix-ui/primitives/pull/3459))
+- Updated dependencies: `@radix-ui/react-primitive@2.1.1`
+
 ## 1.2.3
 
 - Updated dependencies: `@radix-ui/react-presence@1.1.4`