+++ packages/react/menubar/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @radix-ui/react-menubar
 
+## 1.1.15
+
+- Replace deprecated 'ElementRef' with 'ComponentRef' (#3426)
+- Updated dependencies: `@radix-ui/react-roving-focus@1.1.10`, `@radix-ui/react-menu@2.1.15`, `@radix-ui/react-collection@1.1.7`, `@radix-ui/react-primitive@2.1.3`
+
+## 1.1.14
+
+- Updated dependencies: `@radix-ui/react-collection@1.1.6`, `@radix-ui/react-menu@2.1.14`, `@radix-ui/react-primitive@2.1.2`, `@radix-ui/react-roving-focus@1.1.9`
+
+## 1.1.13
+
+- Updated dependencies: `@radix-ui/react-collection@1.1.5`, `@radix-ui/react-menu@2.1.13`, `@radix-ui/react-primitive@2.1.1`, `@radix-ui/react-roving-focus@1.1.8`
+
 ## 1.1.12
 
 - Updated dependencies: `@radix-ui/react-menu@2.1.12`