+++ packages/react/slider/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @radix-ui/react-slider
 
+## 1.3.5
+
+- Replace deprecated 'ElementRef' with 'ComponentRef' (#3426)
+- Updated dependencies: `@radix-ui/react-collection@1.1.7`, `@radix-ui/react-primitive@2.1.3`
+
+## 1.3.4
+
+- Updated dependencies: `@radix-ui/react-collection@1.1.6`, `@radix-ui/react-primitive@2.1.2`
+
+## 1.3.3
+
+- Updated dependencies: `@radix-ui/react-collection@1.1.5`, `@radix-ui/react-primitive@2.1.1`
+
 ## 1.3.2
 
 - Updated dependencies: `@radix-ui/react-use-controllable-state@1.2.2`